# openapi-spec.yaml
openapi: 3.0.1
info:
  title: A swagger API
  description: powered by Flasgger
  termsOfService: /tos
  version: 0.0.1
servers:
- url: /
paths:
...
  /another:
    get:
      summary: A GET endpoint for another resource with multiple query parameters
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: age
        in: query
        required: true
        schema: {}
      responses:
        "200":
          description: A response with query parameters
          content: {}
    put:
      summary: A PUT endpoint for another resource
      responses:
        "200":
          description: A response for another PUT request
          content: {}
...
  /hello/{name}:
    get:
      summary: A simple GET endpoint to return a Hello World message
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          default: World
      responses:
        "200":
          description: A personalized Hello message
          content: {}
    post:
      summary: A simple POST endpoint to return a Hello World message
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          default: World
      responses:
        "200":
          description: A Hello World POST message
          content: {}
...